Keep just the test for FunctionalController
